batch: Group saved batch state modules
The batch package has bounded algorithmic responsibilities while saved
metadata, Git references, lifecycle, and queries remain among flat
operation modules.

Following persistence requires contributors to connect schema validation,
compatibility files, reference publication, content commits, and batch
names across that directory.

This commit groups the complete saved-state responsibility beneath state
and gives reference, name, and compatibility modules paths that describe
their roles.

Later commits will migrate source, file storage, attribution, and command
callers before the temporary state imports are removed.
