+++ src/git_stage_batch/batch/state/__init__.py
@@ -1,0 +1,1 @@
+"""Saved batch metadata, references, lifecycle, and queries."""
